Resolved issues on pull request: - Deleted UserRepository - refactored User equals method - refactored package structure to crm.food.establishment.user
