+++ user-manager-service/src/main/java/com/crm/food/establishment/user/UserManagerApplication.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.usermanager;
+package com.crm.food.establishment.user;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
