+++ user-manager-service/src/main/java/com/crm/food/establishment/user/manager/entity/Role.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.usermanager.crud.entity;
+package com.crm.food.establishment.user.manager.entity;
 
 public enum Role {
     CLIENT,
