+++ user-manager-service/src/test/java/com/crm/food/establishment/user/UserManagerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.usermanager;
+package com.crm.food.establishment.user;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
